Handle situations where the availability api is declined (#6)
* Handle situations where availability API is declined

In some circumstances you'll receive a message:
```
{'message': 'Availability declined'}
```

Let's handle that better, so the user knows what to do.

* Use a session to acquire cookies during API requests

Ryanair now requires a session cookie to be included in queries to the availability API.
Without this, the API now always responds with an "Availability declined" message. We now just need to acquire a session cookie that allows subsequent API requests to work as expected.